requisito 6 corrigido com ajuda do colega Jose Felipe

+++ desafio1.sql
@@ -5,7 +5,7 @@ CREATE DATABASE IF NOT EXISTS SpotifyClone;
 CREATE TABLE SpotifyClone.plano(
       plano_id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
       plano_nome VARCHAR(100) NOT NULL ,
-      preco DOUBLE NOT NULL 
+      preco DECIMAL(3,2) NOT NULL 
   ) engine = InnoDB;
   
  INSERT INTO SpotifyClone.plano (plano_nome, preco)
